Finish

+++ src/app/api/notes/[id]/route.ts
@@ -12,12 +12,36 @@ export const GET = async (request: Request, {params}: {params: {id: string}}) =>
     return NextResponse.json(note, {status: 200});
 }
 
-export const DELETE = async ({params}: {params: {id: string}}) => {
-    const note = await prisma.notes.delete({
-      where: {
-        id: Number(params.id)
-      }
-    });
-  
-    return NextResponse.json(note, {status: 200});
+export const DELETE = async (req: Request, {params}: {params: {id: string}}) => {
+  if (!params || !params.id) {
+    return NextResponse.json({ error: 'Missing id parameter' }, {status: 400});
   }
+
+  const note = await prisma.notes.delete({
+    where: {
+      id: Number(params.id)
+    }
+  });
+
+  return NextResponse.json(note, {status: 200});
+}
+
+export const PATCH = async (req: Request, {params}: {params: {id: string}}) => {
+  if (!params || !params.id) {
+    return NextResponse.json({ error: 'Missing id parameter' }, {status: 400});
+  }
+
+  const body = await req.json();
+
+  const updatedNote = await prisma.notes.update({
+    where: {
+      id: Number(params.id)
+    },
+    data: {
+      title: body.title,
+      body: body.body,
+    },
+  });
+
+  return NextResponse.json(updatedNote, {status: 200});
+}
